Rename sources package to provider
