+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/CsvFileSource.java
@@ -8,7 +8,7 @@
  * http://www.eclipse.org/legal/epl-v10.html
  */
 
-package org.junit.jupiter.params.sources;
+package org.junit.jupiter.params.provider;
 
 import java.lang.annotation.Documented;
 import java.lang.annotation.ElementType;
@@ -21,14 +21,15 @@ import org.junit.jupiter.params.ArgumentsSource;
 @Target({ ElementType.ANNOTATION_TYPE, ElementType.METHOD })
 @Retention(RetentionPolicy.RUNTIME)
 @Documented
-@ArgumentsSource(EnumArgumentsProvider.class)
-public @interface EnumSource {
+@ArgumentsSource(CsvFileArgumentsProvider.class)
+public @interface CsvFileSource {
 
-	Class<? extends Enum<?>> value();
+	String path();
 
-	/**
-	 * @return names of enum constants to provide; if no name is given, all declared enum constants are provided.
-	 */
-	String[] names() default {};
+	String encoding() default "UTF-8";
+
+	String lineSeparator() default "\n";
+
+	char delimiter() default ',';
 
 }
