+++ junit-jupiter-params/src/main/java/org/junit/jupiter/params/provider/package-info.java
@@ -3,4 +3,4 @@
  * corresponding {@link org.junit.jupiter.params.ArgumentsSource} annotations.
  */
 
-package org.junit.jupiter.params.sources;
+package org.junit.jupiter.params.provider;
