+++ junit-jupiter-params/src/test/java/org/junit/jupiter/params/ParamsApiPlayground.java
@@ -18,11 +18,11 @@ import java.util.stream.Stream;
 
 import org.junit.jupiter.api.extension.ContainerExtensionContext;
 import org.junit.jupiter.params.converter.JavaTimeConversionPattern;
-import org.junit.jupiter.params.sources.CsvFileSource;
-import org.junit.jupiter.params.sources.CsvSource;
-import org.junit.jupiter.params.sources.EnumSource;
-import org.junit.jupiter.params.sources.MethodSource;
-import org.junit.jupiter.params.sources.ValueSource;
+import org.junit.jupiter.params.provider.CsvFileSource;
+import org.junit.jupiter.params.provider.CsvSource;
+import org.junit.jupiter.params.provider.EnumSource;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.junit.jupiter.params.support.ObjectArrayArguments;
 
 class ParamsApiPlayground {
